removed geo helpers and made a shared batching module
