+++ src/tiledb/cloud/geospatial/__init__.py
@@ -1,3 +1,5 @@
+from tiledb.cloud.geospatial.ingestion import GeoExtents
+from tiledb.cloud.geospatial.ingestion import get_metadata
 from tiledb.cloud.geospatial.ingestion import ingest
 
-__all__ = ingest
+__all__ = (ingest, get_metadata, GeoExtents)
